server/requirements.txt
flask
flask-cors
gunicorn==20.1.0
google-cloud-storage
google-cloud-translate
APScheduler